Fix wiping of pid file on "gdm already exists" message
Sat Aug 25 10:06:39 2001  George Lebl <jirka@5z.com>

	* daemon/gdm.c, daemon/misc.c:  Fix wiping of pid file on
	  "gdm already exists" message

	* daemon/gdm.c, daemon/gdm-net.c:  Don't unlink files on close.
	  Instead only unlink things on final_cleanup.  This fixes the
	  disappearing socket trick.

	* gui/gdmflexiserver.c:  When not local refuse to run.  This is still
	  not perfect.  But I don't suppose we'll get much better.
	  Unfortunately a new string.  The translators will beat me up.

+++ daemon/misc.c
@@ -72,7 +72,8 @@ gdm_fail (const gchar *format, ...)
 
     g_free (s);
 
-    unlink (GdmPidFile);
+    if (GdmPidFile != NULL)
+	    unlink (GdmPidFile);
     closelog ();
 
     /* Slow down respawning if we're started from init */
